restructuration of dependencies
