+++ unified-report/pom.xml
@@ -39,12 +39,7 @@
 		<module>report-filter</module>
 	</modules>
 	<dependencies>
-		<dependency>
-			<groupId>gov.nist</groupId>
-			<artifactId>hl7-v2-validation</artifactId>
-			<version>1.7.0-SNAPSHOT</version>
-			<optional>true</optional>
-		</dependency>
+		
 		<dependency>
 			<groupId>gov.nist.hit.validation</groupId>
 			<artifactId>hit-xml-validation</artifactId>
