+++ unified-report/validation-proxy/pom.xml
@@ -12,5 +12,11 @@
 			<artifactId>report</artifactId>
 			<version>${parent.version}</version>
 		</dependency>
+		<dependency>
+			<groupId>gov.nist</groupId>
+			<artifactId>hl7-v2-validation</artifactId>
+			<version>1.7.0-SNAPSHOT</version>
+<!--			<optional>true</optional>-->
+		</dependency>
 	</dependencies>
 </project>
